버티컬 라인 병합
