+++ lib/ruler_picker_lib.dart
@@ -3,15 +3,12 @@
 library ruler_picker_lib;
 
 import 'dart:async';
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 part 'ruler_picker/ruler_picker.dart';
 part 'ruler_picker/element/ruler_vertical_line.dart';
 part 'ruler_picker/element/ruler_aligned_long_vertical_line.dart';
 part 'ruler_picker/element/ruler_aligned_short_vertical_line.dart';
-part 'ruler_picker/element/ruler_short_vertical_line.dart';
-part 'ruler_picker/element/ruler_picked_short_vertical_line.dart';
-part 'ruler_picker/element/ruler_long_vertical_line.dart';
-part 'ruler_picker/element/ruler_picked_long_vertical_line.dart';
+part 'ruler_picker/element/vertical_line.dart';
